Merge dev to main after `teal` release (#194)
As teal is released we can now merge the `dev` features to `main`.

The [dev branch was deployed to
test](https://github.com/insightsengineering/teal.gallery/actions/runs/13490845137)
and everything works. The failure is because the `main` branch does not
have the new three apps which will be fixed once we merge.

Note: This does not change any code, we are just merging the `dev`
branch to the `main`

---------

Signed-off-by: Vedha Viyash <49812166+vedhav@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Dawid Kałędkowski <6959016+gogonzo@users.noreply.github.com>
Co-authored-by: Dawid Kałędkowski <dawid.kaledkowski@gmail.com>