243: Update Python App Dependencies  (#246)
This closes issue #243 

I tried to upgrade the teal.reporter and make the minimum changes. When
I was updating it was downgrading other teal.packages, so I tried to
keep all of them to the latest version.

To test this PR follow this steps:

1. Checkout the branch
2. Set your working directory on `teal.gallery/python` folder
3. Restore the renv
4. Configure the python environment if necessary
5. As alternative, comment the code that creates the python environment,
just make sure you have `pip`, `pandas` and `numpy` installed. I had
issues with `reticulate` setting the environment
6. Run the app locally
7. [The variable browser module is not
working](https://github.com/insightsengineering/teal.modules.general/issues/936),
therefore go to the Scatterplot Module and from there see that the
buttons are different and that if the content is added to the report the
title is automated to the module title. Moreover the buttons are located
at the top right corner.

If this is verified the update was successful.

In my case now I see the app 
<img width="923" height="376" alt="image"
src="https://github.com/user-attachments/assets/df4e7330-e4f3-4d6f-a5fa-04d4e8f4527a"
/>

---------

Co-authored-by: osenan <oriol@praenoscere.com>
Co-authored-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: m7pr <marcin.kosinski.mk1@roche.com>